Adding timeline base controller
