+++ ember/app/routes/about/new.js
@@ -1,4 +1,8 @@
 import Ember from "ember";
 import AuthenticatedRouteMixin from 'simple-auth/mixins/authenticated-route-mixin';
 
-export default Ember.Route.extend(AuthenticatedRouteMixin);
+export default Ember.Route.extend(AuthenticatedRouteMixin, {
+  model: function() {
+    return this.store.createRecord("timeline");
+  }
+});
