+++ ember/app/templates/about/-form.hbs
@@ -1,0 +1,24 @@
+<h2>Edit {{title}}</h2>
+<div class="timeline">
+  <div class="new-about-inner">
+    <form {{action "save" on='submit'}}>
+      <p>{{input value=title placeholder="Title"}}</p>
+      <p>{{date-picker date=event_date format="MMMM Do YYYY" valueFormat='YYYY-MM-DD'}}</p>
+      <p>{{textarea value=description placeholder="Body" class="timeline-text-area"}}</p>
+      <p><a href="#" {{action "togglePreview"}}>Toggle Preview</a></p>
+      <label for="is_published">Published:</label>
+      {{view Ember.Select content=published selectionBinding="selectedState" id="is_published" value=is_published}}
+      <ul class="edit-posts-list">
+        <li><a href="#" {{action "destroy"}}>Delete</a></li>
+      </ul>
+      <button type="submit" class="btn">Save</button>
+    </form>
+  </div>
+  <div {{bind-attr class=":timeline-preview :preview openPreview:show:hide"}}>
+    <div class="timeline-card">
+      <h4>{{title}}</h4>
+      {{{format-markdown description}}}
+      <span class="date">{{formattedDate}}</span>
+    </div>
+  </div>
+</div>
