+++ ember/app/templates/about/new.hbs
@@ -1,19 +1,1 @@
-<h2>New Timeline Event</h2>
-<form {{action "saveEvent" on='submit'}} class="form">
-  <p>{{input value=timeline.title placeholder="Title"}}</p>
-  <p>{{date-picker date=timeline.event_date format="MMMM Do YYYY" valueFormat='YYYY-MM-DD'}}</p>
-  <p>{{textarea value=timeline.description placeholder="Body" class="timeline-text-area"}}</p>
-  <p><a href="#" {{action "togglePreview"}}>Toggle Preview</a></p>
-  <label for="is_published">Published:</label>
-  {{view Ember.Select
-    content=published
-    selectionBinding="selectedState"
-    id="is_published"
-  }}
-  <p><button type="submit" class="btn">Save</button></p>
-</form>
-<div {{bind-attr class=":timeline-preview openPreview:show:hide"}}>
-  <h1>{{timeline.title}}</h1>
-
-  {{{format-markdown timeline.description}}}
-</div>
+{{partial "about/form"}}
